requirements.txt
azure-storage-file-datalake
openai
tiktoken
requests
token-count
asyncio
aiohttp